backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-core
langchain-text-splitters
youtube-transcript-api
pydantic-settings
httpx
numpy
